pyproject.toml
[project]
name = "nasa2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiofiles>=25.1.0",
  "aiohttp>=3.13.3",
  "annotated-types>=0.7.0",
  "anyio>=4.11.0",
  "asgiref>=3.11.1",
  "asyncio>=4.0.0",
  "attrs>=25.4.0",
  "backoff>=2.2.1",
  "beautifulsoup4>=4.14.3",
  "boto3>=1.42.63",
  "cachetools>=6.2.1",
  "certifi>=2025.10.5",
  "charset-normalizer>=3.4.4",
  "colorlog>=6.10.1",
  "construct>=2.5.3",
  "distro>=1.9.0",
  "django>=6.0.2",
  "django-cors-headers>=4.9.0",
  "django-redis>=6.0.0",
  "django-restframework>=0.0.1",
  "django-storages>=1.14.6",
  "django-stubs[compatible-mypy]>=5.2.9",
  "djangorestframework>=3.16.1",
  "djangorestframework-simplejwt>=5.5.1",
  "djoser>=2.3.3",
  "docstring-parser>=0.17.0",
  "filetype>=1.2.0",
  "fsspec>=2024.12.0",
  "gliner>=0.2.25",
  "google-ai-generativelanguage>=0.6.15",
  "google-ai-haystack>=5.4.0",
  "google-api-core>=2.25.2",
  "google-api-python-client>=2.186.0",
  "google-auth>=2.42.1",
  "google-auth-httplib2>=0.2.1",
  "google-generativeai>=0.8.5",
  "googleapis-common-protos>=1.71.0",
  "grpcio>=1.76.0",
  "grpcio-status>=1.71.2",
  "h11>=0.16.0",
  "haystack>=0.42",
  "haystack-ai>=2.19.0",
  "haystack-experimental>=0.14.2",
  "httpcore>=1.0.9",
  "httplib2>=0.31.0",
  "httpx>=0.28.1",
  "idna>=3.11",
  "jinja2>=3.1.6",
  "jiter>=0.11.1",
  "json-repair>=0.44.1",
  "jsonschema>=4.25.1",
  "jsonschema-specifications>=2025.9.1",
  "langchain>=1.2.10",
  "langchain-community>=0.4.1",
  "langchain-google>=0.1.1",
  "langchain-google-genai>=4.1.1",
  "langchain-ollama>=1.0.1",
  "lazy-imports>=1.1.0",
  "lxml>=6.0.2",
  "markdown-it-py>=4.0.0",
  "markupsafe>=3.0.3",
  "mdurl>=0.1.2",
  "more-itertools>=10.8.0",
  "neo4j>=5.28.2",
  "neo4j-graphrag>=1.10.1",
  "neo4j-haystack>=2.2.1",
  "neomodel>=6.0.0",
  "networkx>=3.5",
  "numpy>=2.3.4",
  "ollama>=0.6.1",
  "ollama-haystack>=5.3.0",
  "openai>=2.6.1",
  "pandas>=3.0.1",
  "pefile>=2024.8.26",
  "pgvector>=0.4.2",
  "polars>=1.38.1",
  "posthog>=6.7.11",
  "proto-plus>=1.26.1",
  "protobuf>=5.29.5",
  "psycopg2>=2.9.11",
  "psycopg2-binary>=2.9.11",
  "psycopg[pool]>=3.3.3",
  "pyasn1>=0.6.1",
  "pyasn1-modules>=0.4.2",
  "pydantic>=2.12.3",
  "pydantic-core>=2.41.4",
  "pygments>=2.19.2",
  "pyparsing>=3.2.5",
  "pypdf>=6.1.3",
  "python-dateutil>=2.9.0.post0",
  "python-ptrace>=0.9.9",
  "pytz>=2025.2",
  "pyyaml>=6.0.3",
  "redis[hiredis]>=7.3.0",
  "referencing>=0.37.0",
  "requests>=2.32.5",
  "rich>=14.2.0",
  "rpds-py>=0.28.0",
  "rsa>=4.9.1",
  "scipy>=1.16.3",
  "sentence-transformers>=5.1.2",
  "six>=1.17.0",
  "sniffio>=1.3.1",
  "sqlparse>=0.5.5",
  "tenacity>=9.1.2",
  "tqdm>=4.67.1",
  "types-pyyaml>=6.0.12.20250915",
  "typing-extensions>=4.15.0",
  "typing-inspection>=0.4.2",
  "tzdata>=2025.3",
  "uritemplate>=4.2.0",
  "urllib3>=2.5.0",
  "uuid7-standard>=1.1.0",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"
